spotkanie1/requirements.txt
tensorflow==2.13.0
numpy==1.23.5
Pillow==9.5.0
matplotlib==3.7.2
scikit-learn==1.2.2
